Updated tests, removed leftovers, refactor drawer folder structure
