+++ src/components/drawers/hotkeys-dialog.tsx
